refactor: move audio models and voices to environment endpoint
- Add audioModels and voices to EnvironmentController response
- Update frontend EnvironmentConfig interface with Voice and AudioModel types
- Update VoiceConfigService to use environment config instead of separate API calls
- Remove /voices and /audio-models endpoints from AudioController
