+++ server/src/main/java/io/github/mucsi96/learnlanguage/controller/EnvironmentController.java
@@ -8,14 +8,20 @@ import org.springframework.core.env.Environment;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import io.github.mucsi96.learnlanguage.model.AudioModelResponse;
 import io.github.mucsi96.learnlanguage.model.ChatModel;
 import io.github.mucsi96.learnlanguage.model.ImageGenerationModel;
+import io.github.mucsi96.learnlanguage.model.VoiceResponse;
+import io.github.mucsi96.learnlanguage.service.AudioService;
+import io.github.mucsi96.learnlanguage.service.ElevenLabsAudioService;
 import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequiredArgsConstructor
 public class EnvironmentController {
   private final Environment environment;
+  private final AudioService audioService;
+  private final ElevenLabsAudioService elevenLabsAudioService;
 
   @Value("${tenant-id:}")
   private String tenantId;
@@ -36,7 +42,9 @@ public class EnvironmentController {
         Arrays.stream(ChatModel.values())
             .map(model -> new ChatModelInfo(model.getModelName(), model.isPrimary()))
             .toList(),
-        Arrays.stream(ImageGenerationModel.values()).map(ImageGenerationModel::getModelName).toList());
+        Arrays.stream(ImageGenerationModel.values()).map(ImageGenerationModel::getModelName).toList(),
+        audioService.getAvailableModels(),
+        elevenLabsAudioService.getVoices());
   }
 
   public record ChatModelInfo(String modelName, boolean primary) {
@@ -48,6 +56,8 @@ public class EnvironmentController {
       String apiClientId,
       boolean mockAuth,
       List<ChatModelInfo> chatModels,
-      List<String> imageModels) {
+      List<String> imageModels,
+      List<AudioModelResponse> audioModels,
+      List<VoiceResponse> voices) {
   }
 }
